FAB5754 fix event emit scenario on iterators
FAB5754 - Shim Part 2

While testing on composer I found that
having the end event being emitted at the
same time as the data event causes problems
if the data handler invokes an async
function such as deleteState and the
close closes the iterator. To avoid this clash
I removed the emitting of the end event.
data will know if it is the end but to drive
an end event you can just call next on the
 iterator. This doesn’t affect the promise
based alternative using async/await implementation.

Change-Id: I93aa3790b2b31acc74aa7add9a1065ab1ce95046
Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>