FAB-12781 Pipeline Changes
This patch creates a pipeline scripts for both x and z platforms
and archive the build artifacts incase of a build failure.

This change also, send a build notification to RocketChat
(jenkins-robot) channel incase of a merge failure on both x and z.

Docker Container logs will be captured as build artifacts on any failure
and post the same on the jenkins console.

Change-Id: I561f0c578fb4c1b5a5942a25fa3463ef5a192386
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>