[FAB-13426] Fix type definition for protobuf bytes
This patch fixes the type definitions for the values which are from
protobuf's "bytes" fields such as KV.value.

While the type for such values is effectively "ByteBuffer," the type
definition in @types/bytebuffer is not created for Node.js, i.e. the
return type for getBuffer() is ArrayBuffer, not Buffer.
Therefore, this patch sets a special type "ProtoBufBytes" for such
values, which includes a minimum set of methods: toBuffer and toString.

FAB-13426 #done

Change-Id: I8dfe80dee92909a04dabf50728aa2381f3162de6
Signed-off-by: Taku Shimosawa <taku.shimosawa@hal.hitachi.com>