[FAB-16496] Clean up exposed protobuf types
Several of the stub and shim APIs return protobuf
types directly from the protobufjs parser:

Stub:

- getCreator() returns a protobufjs SerializedIdentity object
- getTransient() returns a Map<string, Buffer> that we build now, rather than protobufjs
- getSignedProposal() is crazy, and returns a hand crafted structure similar to a SignedProposal object, but with all of the payloads parsed
- getTxTimestamp() returns a protobufjs Timestamp object
- getBinding() should be OK but probably needs checking
- All calls that return iterators return either a protobufjs KV object, or a protobufjs KeyModification object

Shim:

- success() returns a Response object
- error() returns a Response object

By switching to protobufjs v6, all of these types
have changed. Names are camelCase rather than
underscore_sep_erated, and there are no getters or
setters.

This CR changes the TypeScript definitions to match
the new protobuf formats, and also changes the
structure of the "signed proposal" to match the
style of an actual SignedProposal.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I1d84512142ce57e8d089a901ebdc35434720b55d