[FAB-16527] Use Verdaccio in FV/scenario tests
The current test process packages the npm modules
into tarballs, copies them into the chaincode, and
then adds dependencies to the tarballs.

A local npm repository (Verdaccio) is much more
effective at testing the code as if it was the
"real world". The npm modules can be published to
Verdaccio, and the chaincode build process can
pull them from Verdaccio.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I7c789867bca311afb18e3b52c9ece11954c768ad