[FAB-16602] Fix Node.js chaincode event publishing
The event name was being sent in snake_case, not in
camelCase, so it was being ignored. This change was
incorrectly made recently (by me!) when migrating to
the new protobufjs library.

There are no tests that check chaincode events work,
so I've added a test at the same time so nobody can
do this in the future.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: Ia202c59065587f9bcf36b02c2770f591e8a5c988