[FAB-16714] Add build.sh and start.sh scripts
Currently the build and start processing for Node.js chaincode is
specified in the peer code (platform.go and dockercontroller.go).

We want to move this into the Node.js chaincode code so that we
can update the build and start processing without having to modify
peer code.

To achieve this, we will make it work just like Java chaincode - the
peer will call out to build.sh and start.sh scripts.

Need to add the scripts first, then update the peer code to use them.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I9452f34c128bff597a8af4379103bff92fe49349