[FABCN-397] State queries limited to 100 results
Errow was the new version of protobufjs used to create the bundle.js
gave a different property to the 'has more' field.  hasMore replaced
has_more

As this accessed as JS property it didn't fail, but meant only a limited
number could be accessed.

corrected, added test for 229 results.
bit of dead code pruning

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>