Chaincode errors should be returned as COMPLETED
The Node shim differs from the Go shim in that returns any chaincode generated errors as grpc errors.
In Go, it just returns the ProposalResponse message with a status 500 code and error message embedded in it.
This difference makes it hard to write consistent error handling code in Fabric (esp Gateway) and the SDKs.
This commit changes the Node shim error handler to behave the same as the Go shim.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>