Swap to using the published fabric-protos libraries
Move the codebase to directly using the published protobuf libraries. The fabric-protos
repo, as well as holding the 'main' copy of the protos, also publishes packages for each language
These can then be pulled dirctly here; everything under version control.

Removes the need to the copies of the protobufs held locally in this repo.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>