Recent builds have failued with the browserlist module reporting (on stderr) that it was outdated. This is a dependency of nyc, code coverage.
nyc is 'resting' and not actively being maintained.

Rush version used here is not ignoring that warning, and build fails.

Can't alter the rush rebuild command to ignore warnings. It's really
just going on anything written to stderr.

So adding a redirect on the nyc command stderr-> stdout (2>&1)
Have checked that a drop in coverage does still cause a failure.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>